[FIX] web_studio: reportEditor resets layout from company
Set the external layout on the company. Edit a report that calls
the "web.external_layout" template.

In the report editor, hit reset report and include the header/footer.

Before this commit, the company's layout was not reset.

After this commit it is resetted. The reason was that the external_layout is actually selecting the company's
layout via a dynamic t-call, which is ignored by the code that retrieves views.

opw-4167257

closes odoo/enterprise#69707

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>