[FIX] account_reports: no default tax in tax closing
Some special cases may require to have an account set
as "used in tax closing" and at the same time having
a default tax to improve usability.

This would cause issues as the tax would be set on
the line and trigger the _check_taxes_on_closing_entries
constrains.

This fix aim to mitigate that by ensure that _compute_tax_ids
is not done on tax closing lines.

task id # 3903331

closes odoo/enterprise#69713

Related: odoo/odoo#179612
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>