[FIX] account_report: fix traceback when the user tries to compare reports
Currently, a traceback is occurring when the user tries to compare
periodically in account reports having a fiscal year of date today.

To reproduce this issue:

1) Install Accounting
2) Enable fiscal year from accounting settings
3) Now create a fiscal year record with start and end dates as today
4) Open balance sheet report
5) Filter the comparison as `Previous Period`

Error:- "TypeError: can't compare datetime.datetime to datetime.date"

After the recent changes from this commit, this error is occurring because
`period_date_from` is getting from `fiscal_year`. which is a `datetime.date`.

https://github.com/odoo/enterprise/blob/4c1947963be235f733a56ee3d04b9f8f64d213ea/account_reports/models/account_report.py#L4769-L4775

But here the `comparision_date_from` is of type datetime.datetime,
which leads to a traceback when comparison is done between two.

https://github.com/odoo/enterprise/blob/4c1947963be235f733a56ee3d04b9f8f64d213ea/account_reports/models/account_report.py#L4796-L4797

When the fiscal year record is there we can convert it to datetime.datetime,
By doing this we can resolve this issue.

sentry-5695535237

closes odoo/enterprise#69969

X-original-commit: ea724fb99b16ebdad8ee9b985fa91da928a56a09
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>