[FIX] account_reports: remove parent_id edition
Steps to reproduce issue:

1. Go to _Accounting > Configuration > Account Reports_
2. Select Balance Sheet and Add a line in it
3. Select the new line and give it ASSETS as parent_id
4. Go to Reporting > Balance Sheet
5. The new line is still at the end of the report

Explanation:

Lines have a `sequence` to sort them in the report. It is not computed to change along with `parent_id`, since the lines are supposed to be arranged through the `AccountReportListRenderer` widget. Because of it a line whose `parent_id` changes will not be placed correctly.

Fix reasoning:

The `parent_id` field will be removed from the view, the hierarchy is already shown in the widget.

opw-4101282

closes odoo/enterprise#69568

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>