[IMP] account_reports: get custom display config from root report if no custom handler is set
When we create a tax report that inherits from the generic tax report we do not need always to create a custom handler.
But we might need the custom display config from the root report.
For example all tax reports now have a new filter inside the Date filter for the tax period. We want all reports that inherit from the generic tax report to have that filter.

closes odoo/enterprise#70076

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>