[IMP] test_web_studio: remove useless test helper
There is no need to assert records individually as there is already an
assertXML method, which in the end asserts more information such as the
`noupdate="1"` attribute on some export, or the order of the records
which may be important in some cases. The previous `assertRecords`
method was not testing the records order.

Part-of: odoo/enterprise#69793
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>