[IMP] account_accountant: pdf preview small screen
When being on a small screen, the attachment preview is removed because of a
media query.
To solve that we added a new popout icon that is added when the windows size is
smaller. When clicking on a line and then the popout icon, if the line has an
attachment the pdf is open in an external view.

Also changing the size of the preview viewer, it was a bit small.

closes odoo/enterprise#69824

Task: 4173641
Related: odoo/odoo#179991
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>