[FIX] pos_urban_piper: remove excessive calls to syncAllOrders
The excessive calls on syncAllOrders result to order duplicates in the pos app.
And because of duplicates, the app crashes when "accepting" or "marking as
ready" an urban piper order. The removal of those extra calls results to a more
reliable use of the urban piper features.

closes odoo/enterprise#69778

Related: odoo/odoo#179778
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>