[FIX] project_sale_subscription: allow task creation on non recuring product upsell
Steps to reproduce:
- Download 'Project' and 'Subscription'
- Create a product with:
-- Product type: 'Service'
-- Create on order: 'Task'
-- Project: Pick any
-- Recurring checkbox must not be ticked
- Subscription > New > Add any recurring product > Confirm
- Create invoice > Confirm invoice > Back to subscription
- Upsell > Add your task creation product > Confirm
- The project is linked but no task is created

The ability to create tasks from upsell orders was disabled in
cbbc70f73e6fe9e29e27dea4415e20a262bbdf82.
According to the commit message, this was done because upsell orders
prorate the price of subscriptions accoding to the time until next
invoice relative to the subscription's recurrence (i.e. if the upsell is
created halfway through the recurrence it is discounted by 50% etc...).
Since we do not have a way to prorate tasks it is marked as a technical
limitation.

This however does not need to extend to non-recurring products on
subscription sale orders as they are not prorated.

opw-4114049

closes odoo/enterprise#70972

X-original-commit: 2889a1a6fca3becc03f7b591033a81ad0cc96fcd
Signed-off-by: Yoann Baron (bary) <bary@odoo.com>