[FIX] stock_barcode: Don't update lotName when product isn't track
Currently we only expect a tracking data from a barcode when needed.
However with RFID and sgtin, they need to be unique. So we will
receive a serial number even for product we don't want to track.
So we correct the code to only update the lot if needed.

Part-of: odoo/enterprise#68825
Related: odoo/odoo#177738
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>