[FIX] knowledge: fix bug preventing the share panel from closing
Currently, the share panel cannot be closed once it has been opened. This
commit resolves the issue by preventing the share panel from closing
only when the active element is a modal.

Steps to reproduce the issue:
1. Open Knowledge.
2. Click on the share button to open the share panel.
3. Click on the share button again to close the share panel.
=> The share panel remains open.
TO BE: The share panel should close.

Details:

In the KnowledgeTopbar component, we added external event listeners to
detect when the dropdown menu is opened, closed, or about to close.

When the dropdown is about to close (i.e., when the `hide.bs.dropdown`
event is triggered), we check if the active element is the HTML document.
If the focus is on another element, we prevent the dropdown from closing.
This mechanism was added to prevent the dropdown from closing when users
are adding new members through the invite modal.

Using the useActiveElement hook, the active element can change when a
component is mounted. When the web editor in Knowledge is mounted, the
active element becomes the editor toolbar. As a result, the active
element is no longer the HTML document, and the event listener always
prevents the dropdown from closing.

To resolve this issue, we will now check if the active element is a modal
before preventing the dropdown from closing. This ensures that the dropdown
behaves as expected and the share panel can be closed when necessary.

Reference: https://github.com/odoo/odoo/pull/175880

task-4179706

closes odoo/enterprise#69861

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>