[IMP] account_reports, l10n_*_intrastat, l10n_*_reports, mrp_*, sale_*: handle CTA conversion in reports
The CTA conversion method is used when consolidating multiple companies with different main currencies. The general idea is to use different rates depending on the account we're consolidating:
    - Income/Expense accounts should use the average rate on the considered period
    - Equity accounts should use the historical rate at the date they were made
    - Other accounts should use the current rate, the one that applies at the end of the considered period

Before this commit, Odoo handled such conversions by always applying the current rate everywhere. For this, a structure called the currency table was used.

The community counterpart commit modifies the currency table in order to handle CTA conversion. This commit converts the reports to this new structure, and adds a field on account.report to control whether the report should use 'current' rates or CTA in case of multi-currency consolidation.

We also change the way the comparison feature of the reports works with multi-currency consolidation. In the past, the 'current' rate at the date_to chosen in the reports filter was used for every compared period. Now, each compared period computes a different rate depending on its date bounds. So, the 'current' rate will correspond in each period to the rate applying at the end of the period. The 'average' rate is also computed on each period individually.

Non-accounting reports defined in some sale and mrp modules are also impacted by this commit, to update the way they call the currency table. Their behavior is unaffected, they still make use of the current rates at the time they're opened.

task-4129647

closes odoo/enterprise#69859

Related: odoo/odoo#179914
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>