[FIX] account_online_synchronization: Fix connected until to show reconnect bank button
The aim of this commit is fixing the connected until widget
to show the reconnect bank button when the expiration date is
for today.
The right to fix that, is doing the fix in the view but to avoid
breaking the stable policy, we decide to make the change in the JS
widget.

task-4058051

closes odoo/enterprise#69858

Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>