[IMP] sign: make redirect button label mandatory
This commits makes the redirect button label mandatory when there is
a specified redirect URL. It also improves the hiding of the redirect
label field when there is no redirect URL specified. After this commit,
we changed the default redirect url text from 'Open Link' to 'Close'.
Additionally, when the redirect URL is specified, we no longer show
the Close button for simplifying the UI and not repeat the redirect behavior.

task-4174120

closes odoo/enterprise#74109

X-original-commit: 31c0d596c53b530413ca9d26b6c25af0c65d9818
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>