[FIX] hr_appraisal: prevent error when clicking on Ongoing Appraisal button
When the Appraisal of employee is archived and a user clicks on
the Ongoing Appraisal button, a traceback will appear.

Steps to reproduce the error:
- Install "hr_appraisal" module
- Create new employee > Click Request Appraisal > Save Appraisal >
  Archive that Appraisal
- Go back to employee > Click on "Ongoing Appraisal" button

Error: A traceback appears:
"UnboundLocalError:
local variable 'relevant_appraisals' referenced before assignment'"

https://github.com/odoo/enterprise/blob/a9a3ece5c8ae5fa0b2fa117acb5dbedc69689478/hr_appraisal/models/hr_employee_base.py#L30-L37
When appraisal is archived, "employee_appraisals" will be empty,
So, eventually "relevant_appraisals" will be Empty.
So, it will lead to the above traceback.

sentry-5830582533

closes odoo/enterprise#70289

X-original-commit: e10f7c92e877a540a2a3c8a08c003e4d12ab04ef
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>