[FIX] sale_subscription: revert invoice check on sub renewal
This reverts commit 8bd19ee1dbba0c8dce36a20bdd1e99922084ea24.

This fix prevented a valid use case from functioning: when transferring a subscription to another company, we lose access to the invoices made by the original company. Because of this we would incorrectly prevent the generation of renewal and upsell orders (The subscription was invoiced but invoice_ids does not reflect it).

Other alernatives are also problematic like using invoice_status (which fluctuates for subscriptions) or last_invoice_date (Which could break if the next invoice date is changed manually).

opw-4116538

closes odoo/enterprise#69903

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>