[IMP] account: single lock date per exception
We want to create 1 exception per lock date change (and not bundle
multiple lock date changes in a single exception).
This i.e. allows for more granularity when revoking exceptions.

In the related community PR we changed the lock date exception model.
The change to the model also allows us to (easily) remove a lock date
with an exception (and not just decrease it).

This commit adapts the Lock Dates Wizard to the changes

Related: task-3891414 (lock dates rework)
Related: odoo/odoo#180064
Part-of: odoo/enterprise#69954
Signed-off-by: William André (wan) <wan@odoo.com>