[IMP] l10n_au_hr_payroll_account: Implement Full file replcement of STP
ATO allows a the latest submission for the employees to be
replaced entirely in case of a faulty submission. This allows
the user to correct the YTD values without having to create
an update event seprately. This commit adds the functionality
to replace a file and setting payslips to draft where required.

Task# 3934642

closes odoo/enterprise#65395

Related: odoo/upgrade#6520
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>