[IMP] appointment: add appointment status for gantt
Purpose
=======
Add a status management for gantt view for establisment such
as restaurant to have clearer overview on the bookings and
their status.

Specification
=============
Adding multiple appointment status:

- request: set when the "manual confirmation" option is checked on
the appointment type. Waiting a manual action from the user to confirm the event.
This blocks the slot. The "manual confirmation" can now be set for
appointment based on users as well.
- booked: The slot is blocked and the event is confirmed. This is the default status
when the event has an appointment type.
- attended (Checked-In): Set manually by the user to confirm the presence of
the clients to the event.
- no-show: Set manually by the user to say that the event will not happen.
This will block the slot.
- cancelled: Set manually by the user or automatically set when the event is not active
to say that the event has been cancelled. This will free the slot.

Task-4028357

Part-of: odoo/enterprise#66076
Related: odoo/upgrade#6214
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Co-authored-by: amdi-odoo amdi@odoo.com
Co-authored-by: jeh-odoo jeh@odoo.com