[FIX] knowledge: avoid sidebar entry for virtual record
When the no-content helper is shown and the user clicks on the "New"
button to create a new article, the sidebar lists the newly created
article in the "workspace" and in the "private" section. This is an
undesirable behavior as the newly created article should only appear
in the "private" section.

When no record is passed to the form view, the form view creates a virtual
record. When the user clicks on the "New" button, the system will create
a new article and open it using the `openArticle` function defined in
the component environment.

When calling the `openArticle` function, the function will check if the
article has a title and, if it hasn't, it will fetch the first heading
of the article and use it as title (see: `ensureArticleName`). If the
record is not virtual, the function `openArticle` will then persist the
changes made on the current record before opening the new record.

For its part, the sidebar listens to the changes made on the loaded
record (using a record observer) and updates its state based on the
record changes. When calling the `ensureArticleName`, the title of the
virtual record changes and the sidebar detects that change and adds a
new entry for the virtual record. The sidebar will then add another
entry for the newly created record.

To fix the issue, the `openArticle` function will no longer call the
`ensureArticleName` function when the current record is virtual. The
virtual record state will hence no longer be updated when opening a
new article. As a result, the sidebar will not detect any change and
will no longer create an entry for the virtual record.

Note that there was actually no reason to update the title of the virtual
record as it gets discarded when leaving Knowledge.

Steps to reproduce the issue:
1. Delete all articles from the database
2. Open the Knowledge editor (you should see the no-content helper)
3. Click on the "New" button

=> There is two entries for the new article: one in the "workspace"
   section and one in the "private" section.

TO BE: There should only be one entry in the "private" section.

task-4160386

closes odoo/enterprise#69986

X-original-commit: 9d185d78813f776249a4de0469d659723a94d026
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>