[ADD] account_loans: loans management
Purpose
------
The objective is to compile a comprehensive list of all loans undertaken by the company to maintain a holistic and forecasted view of upcoming due dates (e.g., cash forecast). Odoo will then generate monthly journal entries automatically.
Once the loan record is filled out, the entire process is automated.

A loan can be created in 3 diffrent ways:
- manually by creating each line of the schedule
- by importing the schedulefrom a file (csv or xls for instance)
- by using the compute wizard that will generate the schedule given a couple of parameters

In every case, we'll have a date, a principal, and an interest amount.

Odoo will then handle the rest. We will allow manual entry of the borrowed amount,
the total interest, and the duration to enable control and validation of the accuracy
of the amortization schedule. If the amounts do no match, these will be set in red.

By default, it will be in draft so that the user can configure the journal
of the future generated entries, the long and short term accounts, the expense account,
the asset group (e.g. a house, a car, ...). For more flexibility, the user can
even add its own properties via the cog menu.

When validating the loan, we will use a mechanism a bit similar to the grouped mode
of the deferred mangamenet: for each loan line, we'll generate 3 entries:
- one for the principal and the interest of the due month (the `payment_move`) dated at the end of the month
- one to reclassify the sum of the principals of the next 12 months from the long term account to the short term account (dated on the last day of the month)
- one to reverse the previous reclassification on the first day of the next month

With this mechanism, the LT-ST reclassification is always up to date and
the reco matching can easily be done with the first entry.

The loan can be:
- closed automatically when the last generated entry is posted.
- cancelled manually by the user. In this case, we delete all generated entries (draftor not).
- closed manually by the user. In this case, he/she can enter a date, and we'll delete
all draft entries generated by the loan line whose date is posterior to the chosen date.
- reset to draft

Besides the list and form views, we also provide the graph, kanban and pivot views.

A new report "Loan analysis" is also available under the Report section which shows a Pivot View of all running loans.

task-id 4077613

closes odoo/enterprise#68698

Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>