[FIX] account_invoice_extract: handle guess_country empty array
The stdnum function guess_country can return an empty array if it is not
able to detect a country for a vat number. This fix handles the case of
no country being detected. This can happen for the Swiss vat numbers, as
even though they are valid, stdnum does not check for swiss vat numbers.

closes odoo/enterprise#70013

Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>