[FIX] web_gantt: fix inactive record fetch
Purpose
=======
Fix the inactive records fetching in the gantt view.

Specification
=============
When fetching the records to display in the gantt view from their ids,
we don't specify if the inactive records need to be included in the search.
This is an issue as further down the execution flow, the '_where_calc'
method by default excludes inactive records except if they are explicitly asked for.
=> This means that filtering on inactive records doesn't display any records.

Fixing that by setting the 'active_test' key in the context to 'False' to be
sure we search records with these ids both in active and inactive records.

Related commit: 836909419893b522890776d3ec2f16476ae21108

closes odoo/enterprise#70032

Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>