[FIX] pos_six_iot: handle old and new get config method
As we replaced the individual config files by a single `odoo.conf`, we need to adapt the code to
keep compatibility between updated and non-updated IoT Boxes.

As we did not update this one in the migration method from [https://github.com/odoo/odoo/pull/173866](this PR),
we have to try both old and new methods to get the config value.

closes odoo/enterprise#70067

X-original-commit: 423a226fe19eda30e5eca4ace5a1ca3f805ae50e
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>