[FIX] sign: prevent the error when try to send the template
This error occurs when we delete the user's role in the configuration and then
attempt to 'Send' the template in Templates.

Steps to reproduce:
- Install the 'sign' module
- Sign > Configuration > Roles > Delete 'User' role
- Now go to 'Templates' and Click on the 'Send' and 'Sign Now' button
  of any template

Traceback: ValueError
External ID not found in the system: sign.sign_item_role_user

The error in [1] occurred because the 'sign_item_role_user' ID is missing from
the 'Sign' model, as it was deleted.

[1]- https://github.com/odoo/enterprise/blob/ec873122c01f5312e907a96e462ace794e12fe18/sign/wizard/sign_send_request.py#L132

sentry-5694515390

closes odoo/enterprise#80587

X-original-commit: f2d26c89a038f1dc127f2cd9348b08b9df4063ca
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Aayush Modi (aamo) <aamo@odoo.com>