[IMP] stock_enterprise: add route start & end to google map url
This commit extends the map view to add the warehouse address as a start
and end point for the route generated by google maps in the map view of
stock pickings. The warehouse address will be used only if it was the
same for all stock pickings that are shown in the view. If the shown
pickings had different warehouses, the default behavior is applied,
which is to show the current user location as a start point and no end
point.

Task-4173210

closes odoo/enterprise#70099

Related: odoo/odoo#180327
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>