[FIX] l10n_mx_reports: allow the report coa sat on debit + credit 7X accounts
Steps to reproduce:

1. Install l10n_mx.
2. Activate developer mode.
3. Activate Download XSD files (XML validation).
4. Go to accounting  configuration > chart of accounts
5. Make sure here to have 2 accounts like 702.XXX. One should have
the credit tag and the other one the debit one.
8. Go to reports and select Trial Balance.
9. Click on the download drop list and select COA SAT

Issue: Up to now we have just intentionally marked this accounts
as no valid, we can see that inside trial_balance.py for l10n_mx_reports
at L201 this was a known limitation.

Solution: As we can see from the technical anex in the mexican goverment
site https://www.gob.mx/sat/documentos/contabilidad-en-medios-electronicos-anexo-tecnico
we can see that there are asset, liability and equity accounts which,
by their nature, can be presented as debit or credit.

opw-4185713

[FIX] l10n_mx_reports: Fix date formatting in _l10n_mx_set_options_non_month_13

This makes the options not serializable.

closes odoo/enterprise#82363

X-original-commit: 9bd21c30e000118ef4864d4e2ca6fc636ee615b8
Signed-off-by: Laurent Smet (las) <las@odoo.com>