[IMP] mrp_workorder: allow limiting WO._calc_cost to specific date
Sometimes we only want the total cost of a WO up to a specific date
(e.g. for WIP accounting), therefore we adjust the existing method to
handle both the existing all WO cost and the cost for any WO recorded
time with an end date before the passed date.

task-3952348

closes odoo/enterprise#70252

Related: odoo/odoo#180640
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>