[IMP] l10n_co_dian: send the AttachedDocument by email
Due to legal reasons, we need to send a zip containing the PDF and the
xml. The xml is an UBL `AttachedDocument` containing both the initial
xml representing the invoice and the `ApplicationResponse` returned when
calling the GetStatus webservice of DIAN.

The `ApplicationResponse` is an xml representing the state of the
invoice on the DIAN server.

task-4133688

closes odoo/enterprise#70171

Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>