[FIX] industry_fsm: allow resequence projects in Configuration > Projects
Before this commit, it was not possible to resequence the projects in
any list views of Project and Field Service because the `handle` widget
is not defined in the sequence and the default order
applied in those list views does not begin with `sequence` field.

This commit shows the sequence field in the list view of `Field Service
> Configuration > Projects` and change the default of that list view to
allow the user to resequence the projects as he wishes.

Community PR: odoo/odoo#180403

Close odoo/odoo#176961

closes odoo/enterprise#70137

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>