[FIX] account_reports: Fix account codes appearing as 'False' in reports
Since sharing accounts between companies in odoo#171079, an account's
code may be False from the perspective of the active company. In that
case, we want only the account name to appear in reports.

A fix was already done for `account.display_name` in odoo#177943.

This commit does a similar fix for reports which use code + name rather
than display_name.

closes odoo/enterprise#70157

Taskid: none
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>