[FIX] account_intrastat: floating point issue in intrastat report
With an EU company (i.e. BE)
Open a product [PROD], in Accounting Tab set:
- Commodity Code: 84716070
- Supplementary units: 0.56
Create an invoice to an EU partner (i.e. IT) with 5 units of [PROD]
Confirm invoice
Accounting > Reporting > Intrastat Report
Export pdf or xml

Issue: Report will show a very large decimal number
This occurs because the number is computed via SL query that is not
rounded

opw-4160104

closes odoo/enterprise#70321

X-original-commit: b0ce3df1180c3f074fc84a0e48e7c43724281f61
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>