[FIX] website_knowledge: fix hidden public view on desktop
Currently, the article content appears empty on desktop devices because
the main container (`.o_knowledge_public_view`) no longer occupies the
full vertical space of the page.

That issue was introduced by [1].

To resolve this issue, we will remove the wrapper elements: `div#wrapwrap`
and the `main` element from the generated pages. This will simplify the
DOM structure and allow the main container to take up all the available
vertical space without needing to set the height for all containers.

Steps to reproduce the issue:
1. Open Knowledge
2. Publish an article using the share panel (click on the "Share" button)
3. Copy the share link (`/knowledge/article/<article_id>`​)
4. Open an incognito tab and access the link copied at step 2

=> The article content is hidden on desktop.

TO BE: The article content should be visible.

[1]: https://github.com/odoo/odoo/pull/98429

task-4193478

Part-of: odoo/enterprise#70211
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>