[PERF] documents_spreadsheet: use thumbnail checksum
The Documents kanban view is slow when the workspace contains large
spreadsheets.

In the Documents app, the `checksum` field can't be used has a "unique" value
to invalidate browser cache when it changes. The reason is that the document
attachment isn't actually changed when a new spreadsheet revision is created

The fix was to use the last revision uuid as this unique value.
It has one major drawback though: if there is no active revision, the
revision uuid is read from the snapshot, which is a json file which can be
very large (tens or even hundreds of Mb).
(see a055dfeb84abc34015b269422264534d97a83d8f)

This is very slow.

After this commit, we use the checksum of the thumbnail.

Open the Documents kanban with 8 relatively larges spreadsheets:
- before: ~1.4s
- after: ~35ms (5ms to compute `spreadsheet_thumbnail_checksum`)

An alternative to the proposed solution here is to make `current_revision_uuid`
faster by looking at inactive revisions instead of loading the json.
But
- the time complexity would grow linearly with the number of revisions
- if there's no revisions at all (which happens when uploading a new spreadsheet file),
  the json still needs to be loaded.
This idea is valid on its own, but is left for another commit, the field
`current_revision_uuid` is not used in many situations anyway.
Let's make the Documents kanban view faster in all cases.

closes odoo/enterprise#70217

Task: 4193350
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>