[IMP] website_knowledge: add HTML semantic for public view
This commit adds semantic HTML to the public view of the Knowledge module.
It introduces a header and a main element, which should help web crawlers
index the relevant parts of the article.

These tags have to be reintroduced since the website_knowledge view overrides
the contents of the frontend_layout body.

task-4193478

Part-of: odoo/enterprise#70211
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>