[ADD] hr_recruitment_integration_base: Recruitment platform posting integration
This commit adds a module for recruitment platform posting integration.
This module provides a base for integration with recruitment platforms like
Monster, LinkedIn, etc. In addition to that it also serves as a base for
potential pluggins for other modules of odoo like website

task-4017067

Part-of: odoo/enterprise#70213
Related: odoo/odoo#171382
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>