[FIX] knowledge: fix knowledge_cover_selector_tour
In this commit, we fix the knowledge_cover_selector_tour tour.
Actually, the step 41 makes no sense.
{
    // Make the add cover button visible
    trigger: '.o_knowledge_edit_cover_buttons',
},
We check for edit cover buttons but the cover was removed in previous
steps.
we take advantage of the commit to transform the comments into "content"
to make it easier to debug the tour.
We take advantage of this commit to lint the file.

closes odoo/enterprise#70235

Signed-off-by: Bastien Fafchamps (bafa) <bafa@odoo.com>