[FIX] account_batch_payment: add missing dependency to constraint
payment_method_id is readonly in account.batch.payment's form view when there is at least one payment on the batch. So, it cannot easily be changed in standard in order to break the constraint. It could however happen in custom views or when generating records directly from the code.

closes odoo/enterprise#70256

X-original-commit: 46e8f37a7b370a2f8d62f97459be9e16680959ae
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>