[FIX] planning, sale_planning: regenerate recurring slots from origin
Steps to reproduce:
- Planning > Configuration > Employees > New
- Schedule > By resource > Filter for your employee
- Click on the employee's Gantt line and assign a shift
- Click the shift > Edit > Click the looping arrow next to date
- Your shift should now be recurring over the coming weeks
- Go to the 2nd week > Edit the shift's end date
- Check 'All shifts' at the top and save

There are 2 issues here:
1. A MissingError is raised
2. Even when bypassing that error, the planning is regenerating the
slots from the original slot's start date to the 2nd slot's end date.
(i.e. for a weekly recurrence you will now have 2 week long slots
instead of the expected 1 week)

Both happen because when editing the start/end date of a recurring shift
we simply unlink the existing ones and regenerate them from the original
slot. Since they are unlinked we can no longer access the shift we were
trying to edit (the 2nd occurrence) resulting in the MissingError. And
because we regenerate from the first slot despite taking values from the
second, we stretch the original slot, which is repercuted on every
subsequently generated slot's dates.

opw-4154647

closes odoo/enterprise#71051

X-original-commit: bfd5866ef689d28bb2ef0828f4454d6e7f4d6e3d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Yoann Baron (bary) <bary@odoo.com>