[IMP] knowledge: fix unloaded CSS print rules
When the user presses `Ctrl` + `p` to print an article, the page appears
blank because the assets for print media queries are only loaded after
the user clicks the "Export" button.

To ensure that the print-specific styles are loaded, we will move the
`knowledge_print.scss` file to the `web.assets_web_print` asset bundle.
The generated stylesheet will then be linked to the page using a `link`
tag with the `media="print"` attribute. The browser will then be able
to load the stylesheet when the user is printing the page or previewing
it for print.

Steps to reproduce the issue:
1. Go to the Knowledge app.
2. Open an article
3. Press `Ctrl` + `p` to print the article.
=> The print rules are not loaded and the page appears blank.

TO BE: The print rules should be loaded.

task-4191390

closes odoo/enterprise#72108

X-original-commit: 3b03f2691dbe80bceb973decf650679a2bff9747
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>