[FIX] industry_fsm: hide project in form view if only 1 fsm project exists
Before this commit, the project in the form view is no longer hidden in
mobile in My Tasks of Field Services due to some changes in the form
view of tasks.

This commit updates the form view to make sure to project field is
correctly hidden in mobile when there is only one fsm project.
This commit also adds a placeholder to partner_phone to make sure the
user knows which field it is in mobile view since its label is not
displayed in mobile view.

task-4189791

Part-of: odoo/enterprise#70353
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>