[FIX] account_asset: convert assets non-deductible tax value into company currency
Problem:
When calculating the non_deduictible_tax_value on an asset the non_deductible_tax_value of the original move lines are not converted into the asset's/company's currency. This is used when calculating the original_value of the asset and can cause incorrect calculations.

Purpose of this PR:
To make sure the non_deductible_tax_value is stored in the asset's currency

How to reproduce:
1- install: Uruguay - Accounting
2 - change company country to Uruguay
3- settings -> accounting -> taxes -> Fiscal Country: Uruguay.
4- create a new tax and make sure that the accounts on % of tax are expenses account so that the tax in non-deductible tax.
5- make sure to select the right tax group and country = Uruguay in the advanced options of tax.
6- create new product, and in accounting tap set the fixed asset account to the expenses account.
and in the purchase tap make sure that the vendor taxes is set to the tax created before.
7- create a new vendor bill with the currency set to UYU. and confirm the bill. make sure that the fixed assets account is selected.
8- go to assets and create a new asset, in bills tab add the created bill. the asset will be created and the Original Value filed will contain the wrong amount.
When calculating the non_deductible_tax_value for an asset, if there are original move lines in a foreign currency they won't be converted into the assets currency

opw-4089386

closes odoo/enterprise#72433

X-original-commit: c683467dcd536d6bb2b883125d40fc10f92aa812
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>
Signed-off-by: Emeka Ubakanma (emub) <emub@odoo.com>