[FIX] hr_referral: fix traceback when refusing an application
Steps:
- Install hr_referral, hr_contract_salary
- click on refuse button in application form

Description of the issue/feature this PR addresses:
In the recruitment module, clicking the refuse button on an application results
in a traceback error.

Cause
The issue arises due to the name field, which was removed from the hr_applicant
model (as per https://github.com/odoo/odoo/pull/173288). The system still
attempts to retrieve this field, causing the error.

Fix:
This PR resolves the issue by removing the unnecessary name field from the
applicant model

task-4195716

closes odoo/enterprise#70278

closes odoo/enterprise#74158

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>