[FIX] industry_fsm: allow assignee access to internal only tasks
Steps to reproduce:
- Project > New > Create Project
- Project > ':' menu > Settings > Settings tab
- Set 'Visibility' to 'Invited Internal Users'
- General Settings > Users & Companies > Users
- Create a task and assign it to Marc Demo
- Edit Marc Demo to have 'User' access rights in 'Project'
- As Marc Demo: Project > Tasks > My Tasks > Click your task

Access is denied because Marc is not a follower of the project, and
therefore cannot read the task's 'is_fsm' field (which is computed from
the project). Due to the specifications of the visibility mode, the task
should however be accessible to him.

A fix was previously made for a similar issue in Field Service
5fc4f7d15e56d2b4335aa70e896fdfbd901d6076, where we changed the
_search_is_fsm function to sudo the project for the same reason as here
(accessing project_id.is_fsm) so this should not raise any additional
security concerns.

In that same fix we switched from a related field to a compute,
compute_sudo was however defaulting to False since is_fsm is not stored.

opw-4176243

closes odoo/enterprise#70293

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>