[FIX] Assign journal with transactions and no currency
Commit https://github.com/odoo/enterprise/commit/a47e68a18fa478341efe21d82fa7f0e3e86e4c2c
added an user error when trying to connect a bank journal with an online account
that does not share the same currency, if there are already some moves created
in that journal.

The field currency_id being not required on the journal model, this error would
also trigger when the currency_id was not set on the journal, but it contained
only moves of the same currency as the account that we are trying to connect.

An additionnal condition was added to the check, being that the journal has no
currency set but the moves in that journal all use the same currency than the
online account.
If this is the case, the currency of the journal is set to the one from the
online account.

opw-4177736

closes odoo/enterprise#70565

X-original-commit: 911b03f3c020924b8e917cbd0099a1c773c6ed7a
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Florent Huylenbroeck (flhu) <flhu@odoo.com>