[FIX] appointment: fix resource events display
Purpose
=======
Fix the resource events which weren't displayed in the gantt view
when booked from the front-end.

Specification
=============
When an appointment based on resources is booked from the front-end, the event
"user_id" field is set to be OdooBot.
This is an issue as the read_group domain for the calendar.event model checks
the user_id 'res_users_settings_id' field for privacy purpose but OdooBot
doesn't have any users settings.
=> The read_group domain is always evaluated to False. No resource based events
booked from the front-end are displayed on the gantt view.

Fixing that by updating the privacy domain to be sure every events related to
a resource can now be visible and accessible on the gantt view.
The privacy field of an event is mainly related to the user settings but
resource events aren't typically linked to any user meaning their
visibility shouldn't depend on the privacy.

Task-4191434

closes odoo/enterprise#70182

Related: odoo/odoo#180501
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>