[IMP] industry_fsm{_report}: make sure to keep the right context for pager
Before this commit, when the user is in mobile and go to field service
app and click on a task in the kanban view, he cannot use the pager in
the form view, because a new action is called and the action does not
have the active_ids, that is the records displayed in the kanban view.

This commit makes sure the action called to open a simplified form view
in mobile get the active_ids to be able to use the pager.

task-4189791

Part-of: odoo/enterprise#70353
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>