[FIX] gantt: popovers open in view port
Before this commit, click on a long pill with ends outside of the view
port would sometimes make the gantt popover (that displays some info on
the pill) open outside of the view port. Here we make it target the
middle of the visible part of the pill when that happens.

Task ID: 3989911

closes odoo/enterprise#70447

X-original-commit: 74aeaf280cea2e5998240e28403e43b6f407965a
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>