[FIX] sign: restore sign placeholder field formatting
Steps to reproduce:
- Sign > Edit any template > Add a field
- Click your field > Select left alignment > Validate

The placeholder text overlaps with the arrow block, in 16.0 this does
not occur. The culprit is the ps-0 class added to signItem which removes
all padding from the block.

The formatting was changed in f1b35ae9030ff82afc498bfbd6315414407bb4e1
to make it consistent between the other views (The shift is kept even
the move icon is not visible) but the alignment feels consistent even
without this class.

opw-4159676

closes odoo/enterprise#71098

X-original-commit: a7bd8323b35fed12651619a3aafb19b20135cd06
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Yoann Baron (bary) <bary@odoo.com>