[IMP] currency_rate_live: We are place the date of the day to which the quote applies
This commit addresses the issue of applying incorrect exchange rates for foreign currency transactions.
Previously, we were using the rate from the last closing date instead of the rate applicable to the transaction day.
Moving forward, the correct exchange rate will be used based on the day to which the quote
applies, as outlined in the official guidelines for foreign currency operations by the Dirección General Impositiva (DGI)

closes odoo/enterprise#72104

Latam-task: 1264
Adhoc-side-task: 43762
X-original-commit: 746f3b96f27f117cec05cb4c4957ad5abf2296b4
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>