[FIX] l10n_co_edi: restrict debit_note changes
Steps to reproduce the issue:

1. Activate Colombian localization
2. Create an Invoice and Confirm
3. Go to _Accounting > Configuration > Journals_
4. Select Customer Invoices, set Nota de Débito to True and Save
5. Go back to your Invoice, its Electronic Invoice Type is now "Nota Débito"
6. Set Nota de Débito to False in the Customer Invoices Journal
7. Return once again to you Invoice, its Electronic Invoice Type did not change

Explanation:

When the user changes `l10n_co_edi_debit_note` in an `account.journal`, it changes this value for all entries in said `account.journal`, not only the ones in draft state.
https://github.com/odoo/enterprise/blob/9137e5d9f2690bfe375436c83356ebaa4bd7e8c6/l10n_co_edi/models/account_invoice.py#L65
`_compute_l10n_co_edi_type` depending on `l10n_co_edi_debit_note`, it is called and, if the change is from False to True, `l10n_co_edi_type` is changed to a specific value.
https://github.com/odoo/enterprise/blob/9137e5d9f2690bfe375436c83356ebaa4bd7e8c6/l10n_co_edi/models/account_invoice.py#L72-L81
If `l10n_co_edi_debit_note` is set back to False, `l10n_co_edi_type` is not reverted to its previous value and the user can not manually change this field for posted or cancelled entries.

Fix reasoning:

To avoid such situation from happening, the field will be restricted for journals with entries in a non draft state.

opw-4174980

closes odoo/enterprise#70425

X-original-commit: 0b331ced6c11c87e0a9bb920ecf131c52f6b83bb
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>
Signed-off-by: Stroobant Paul (stpa) <stpa@odoo.com>