[FIX] stock_barcode: avoid missing product RFID
When a SGTIN RFID is scanned, it's linked to a product and a serial
number. If the linked product doesn't exist in the DB, it could happen
the SN is processed as the SN for another product (if the previous
scanned product is also tracked by SN.)
To avoid that, barcode linked to a non-existing product are skipped.

Part-of: odoo/enterprise#68825
Related: odoo/odoo#177738
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>